Use the modern FFI.set_source() method to do out-of-line API mode.
This lets the CFFI backend be used on CPython and is more maintainable
going forward. Initial benchmarking of the CFFI backend on CPython
shows it to be a few percent slower than the cython backend, but
probably not significantly so.

Note that the tests are not expected to pass on Travis/pypy yet,
because Travis pypy is too old. Use of pyenv may be required.
